+++ hydragnn/utils/distributed.py
@@ -14,7 +14,6 @@ import re
 
 import torch
 import torch.distributed as dist
-from torch.utils.tensorboard import SummaryWriter
 
 from .print_utils import print_distributed
 
@@ -170,13 +169,6 @@ def is_model_distributed(model):
     return isinstance(model, torch.nn.parallel.distributed.DistributedDataParallel)
 
 
-def get_model_or_module(model):
-    if is_model_distributed(model):
-        return model.module
-    else:
-        return model
-
-
 def get_distributed_model(model, verbosity=0):
     device_name, device = get_device(verbosity)
     if dist.is_initialized():
@@ -187,30 +179,3 @@ def get_distributed_model(model, verbosity=0):
                 model, device_ids=[device]
             )
     return model
-
-
-def save_model(model, name, path="./logs/"):
-    _, world_rank = get_comm_size_and_rank()
-    if world_rank == 0:
-        model = get_model_or_module(model)
-        path_name = os.path.join(path, name, name + ".pk")
-        torch.save(model.state_dict(), path_name)
-
-
-def get_summary_writer(name, path="./logs/"):
-    _, world_rank = get_comm_size_and_rank()
-    if world_rank == 0:
-        path_name = os.path.join(path, name)
-        writer = SummaryWriter(path_name)
-
-
-def load_existing_model_config(model, config, path="./logs/"):
-    if "continue" in config and config["continue"]:
-        model_name = config["startfrom"]
-        load_existing_model(model, model_name, path)
-
-
-def load_existing_model(model, model_name, path="./logs/"):
-    path_name = os.path.join(path, model_name, model_name + ".pk")
-    state_dict = torch.load(path_name, map_location="cpu")
-    model.load_state_dict(state_dict)
